Cast input_ids to int to pass tests

+++ src/transformers/modeling_tf_t5.py
@@ -782,7 +782,7 @@ class TFT5PreTrainedModel(TFPreTrainedModel):
             decoder_start_token_id is not None
         ), "self.model.config.decoder_start_token_id has to be defined. In TF T5 it is usually set to the pad_token_id. See T5 docs for more information"
 
-        shifted_input_ids = tf.roll(input_ids, 1, axis=-1)
+        shifted_input_ids = tf.cast(tf.roll(input_ids, 1, axis=-1), tf.int32)
         start_tokens = tf.fill([shape_list(input_ids)[0], 1], pad_token_id)
         shifted_input_ids = tf.concat([start_tokens, shifted_input_ids[:, 1:]], axis=-1)
